requirements.txt
Django==2.0.8
djangorestframework==3.8.2
packaging==16.8
pytz==2018.5
